Look up Doom objects by sprite number.
Lots of the symbolic names used in the Doom source for object types
are pretty meaningless (MT_MISC0). By contrast, the sprite names are
all meaningful and are actually used in P_TouchSpecialThing() for
deciding what powerup gets applied on pickup. This approach can't
work for every object type but most powerup items which can be
picked up have a 1:1 mapping between object type and sprite.