Map Weredragon to Imp.
Previously this was mapped to the Baron, but this doesn't work so well:
levels with lots of Weredragons seem very imposing, even though the hit
point count level is adjusted appropriately. The imp fits better since
the color is right (brown) and in gameplay terms the imp and baron are
very similar anyway.