Fix key lamp color mappings.
The lamps are used in Heretic to indicate which key color a door
corresponds to, so it's important these match the door types. These
were previously not consistent with the key color remappings; since
there is no yellow torch, use the yellow candelabra as a substitute
(and switch the current candelabra item to tech lamp).